+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project from version 1.2.0 upwards are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+
+### Added
+- `Node.transformChildren` to run a function on all the children of a Node
+- `Node.replaceWith` to replace a Node with another
+
 ## [1.6.29] – 2024-07-09
 
 ### Fixed